// apps/mobile/http/generated/models/GetOrganizationsSlugCommissionsInstallments.ts
export const getOrganizationsSlugCommissionsInstallmentsQueryParamsDirectionEnum =
  {
    INCOME: "INCOME",
    OUTCOME: "OUTCOME",
  } as const;

export type GetOrganizationsSlugCommissionsInstallmentsQueryParamsDirectionEnumKey =
  (typeof getOrganizationsSlugCommissionsInstallmentsQueryParamsDirectionEnum)[keyof typeof getOrganizationsSlugCommissionsInstallmentsQueryParamsDirectionEnum];

export const getOrganizationsSlugCommissionsInstallmentsQueryParamsStatusEnum =
  {
    ALL: "ALL",
    PENDING: "PENDING",
    PAID: "PAID",
    CANCELED: "CANCELED",
  } as const;

export type GetOrganizationsSlugCommissionsInstallmentsQueryParamsStatusEnumKey =
  (typeof getOrganizationsSlugCommissionsInstallmentsQueryParamsStatusEnum)[keyof typeof getOrganizationsSlugCommissionsInstallmentsQueryParamsStatusEnum];

export type GetOrganizationsSlugCommissionsInstallmentsQueryParams = {
  /**
   * @minLength 1
   * @maxLength 9007199254740991
   * @default 1
   * @type integer | undefined
   */
  page?: number;
  /**
   * @minLength 1
   * @maxLength 100
   * @default 20
   * @type integer | undefined
   */
  pageSize?: number;
  /**
   * @default ""
   * @type string | undefined
   */
  q?: string;
  /**
   * @type string | undefined, uuid
   */
  productId?: string;
  /**
   * @type string | undefined
   */
  direction?: GetOrganizationsSlugCommissionsInstallmentsQueryParamsDirectionEnumKey;
  /**
   * @default "ALL"
   * @type string | undefined
   */
  status?: GetOrganizationsSlugCommissionsInstallmentsQueryParamsStatusEnumKey;
  /**
   * @pattern ^\d{4}-\d{2}-\d{2}$
   * @type string | undefined
   */
  expectedFrom?: string;
  /**
   * @pattern ^\d{4}-\d{2}-\d{2}$
   * @type string | undefined
   */
  expectedTo?: string;
};